Updating archiving of config files to include systemd and non-systemd config file locations.
Also added the "-y" launch.py command line option in order to specifiy
a non-temporary location for the modified test yaml files.

+++ src/tests/ftest/get_remote_files.sh
@@ -153,7 +153,8 @@ scp_files() {
         archive_name="${file_name%%.*}.$(hostname -s).${file_name#*.}"
         if scp -r "${file}" "${2}"/"${archive_name}"; then
             copied+=("${file}")
-            if [[ ! "${file}" =~ test.cov ]]; then
+            if [[ ! "${file}" =~ /etc/daos/ ]] && \
+               [[ ! "${file}" =~ test.cov ]]; then
                 if ! rm -fr "${file}"; then
                     echo "  Error removing ${file}"
                     echo "    $(ls -al ${file})"
